Add files via upload
This is the DecisionTree Model that I have created. I dropped date because it wasn't working well and then I hot encoded a couple of the columns. I also binned the homespread and used that as the y variable (homespread as in home elo - away elo). I'll continue to look into improving this but the classifier did run.